[CDSR-1729] move /scheduled/scheduled-document-upload/choose-files and /scheduled/scheduled-document-upload/continue
